requirements.txt
flask
pandas
python-amazon-paapi
numpy
matplotlib
scikit-learn
openpyxl